+++ plugins/org.preesm.ui.pisdf/src/org/preesm/ui/pisdf/features/SetActorRefinementFeature.java
@@ -155,7 +155,7 @@ public class SetActorRefinementFeature extends AbstractCustomFeature {
       if (bo instanceof Actor) {
         final Actor actor = (Actor) bo;
 
-        final String question = "Please select a valid file\n(.idl, .h or .pi)";
+        final String question = "Please select a valid refinement file (.h or .pi)";
         final String dialogTitle = "Select a refinement file";
         final IPath path = askRefinement(question, dialogTitle);
         if (path != null) {
@@ -185,9 +185,8 @@ public class SetActorRefinementFeature extends AbstractCustomFeature {
     // .idl prototypes
     final Set<String> fileExtensions = new LinkedHashSet<>();
     fileExtensions.add("pi");
-    fileExtensions.add("idl");
     fileExtensions.add("h");
-    return FileUtils.browseFiles(PlatformUI.getWorkbench().getActiveWorkbenchWindow().getShell(), dialogTitle,
+    return FileUtils.browseFiles(PlatformUI.getWorkbench().getActiveWorkbenchWindow().getShell(), dialogTitle, question,
         fileExtensions);
   }
 
@@ -243,17 +242,15 @@ public class SetActorRefinementFeature extends AbstractCustomFeature {
           final List<FunctionPrototype> allInitPrototypes = getPrototypes(file, actor, PrototypeFilter.INIT);
 
           FunctionPrototype initProto = null;
-          if (!initPrototypes.isEmpty() || !allInitPrototypes.isEmpty()) {
-            title = "Init Function Selection";
-            message = "Select an optionnal init function for actor " + actor.getName()
-                + ", or click Cancel\n(* = any string, ? = any char):";
-            final FunctionPrototype[] initProtoArray = initPrototypes
-                .toArray(new FunctionPrototype[initPrototypes.size()]);
-            final FunctionPrototype[] allInitProtoArray = allInitPrototypes
-                .toArray(new FunctionPrototype[allInitPrototypes.size()]);
-            initProto = PiMMUtil.selectFunction(initProtoArray, allInitProtoArray, title, message, false);
-
-          }
+          title = "Init Function Selection";
+          message = "Select an optionnal init function for actor " + actor.getName()
+              + ", or click Cancel to set none.\nNote: prototypes with pointers or arrays as "
+              + "arguments are filtered out.\n(* = any string, ? = any char):";
+          final FunctionPrototype[] initProtoArray = initPrototypes
+              .toArray(new FunctionPrototype[initPrototypes.size()]);
+          final FunctionPrototype[] allInitProtoArray = allInitPrototypes
+              .toArray(new FunctionPrototype[allInitPrototypes.size()]);
+          initProto = PiMMUtil.selectFunction(initProtoArray, allInitProtoArray, title, message, false);
           if ((loopProto != null) || (initProto != null)) {
             this.hasDoneChanges = true;
             final CHeaderRefinement newRefinement = PiMMUserFactory.instance.createCHeaderRefinement();