+++ release_notes.md
@@ -32,6 +32,7 @@ PREESM Changelog
 * Fix #121
 * Fix #77
 * Fix #87
+* Fix #155: always show the init function selection dialog and add a note explaining which functions are filtered out;
 
 
 ## Release version 3.8.1